Attempted to compile with C++23 (2)
header guard, sprintf_s, double to float implicit conversion, enum C-style casts

+++ physx/source/physxcharacterkinematic/src/CctCharacterController.cpp
@@ -776,7 +776,7 @@ static SweepFunc gSweepMap[SweptVolumeType::eLAST][TouchedGeomType::eLAST] = {
 	}
 };
 
-PX_COMPILE_TIME_ASSERT(sizeof(gSweepMap)==SweptVolumeType::eLAST*TouchedGeomType::eLAST*sizeof(SweepFunc));
+PX_COMPILE_TIME_ASSERT(sizeof(gSweepMap) == static_cast<int>(SweptVolumeType::eLAST) * static_cast<int>(TouchedGeomType::eLAST) * sizeof(SweepFunc));
 
 static const PxU32 GeomSizes[] =
 {
